--- cypress.env.json
@@ -1,4 +1,0 @@
-{
-  "qase_email": "nikolanikic11@gmail.com",
-  "qase_password": "VTS2023Test"
-}
