+++ cypress/support/swag-labs/loginPage.js
